+++ .formatter.exs
@@ -9,7 +9,8 @@ locals_without_parens = [
   slot: 2,
   catalogue_test: 1,
   catalogue_test: 2,
-  load_asset: 2
+  load_asset: 2,
+  embed_sface: 1
 ]
 
 [