+++ test/surface/components/live_redirect_test.exs
@@ -93,7 +93,7 @@ defmodule Surface.Components.LiveRedirectTest do
         class: "link",
         method: :delete,
         data: [confirm: "Really?"],
-        csrf_token: "token"
+        "csrf-token": "token"
       )
 
     assert attr_map(html) == attr_map(actual)