+++ test/surface/macro_component_test.exs
@@ -234,21 +234,48 @@ defmodule Surface.MacroComponentTest do
     assert output =~ ~r"cannot render <#MissingExpand> \(MacroComponents must export an expand/3 function\)"
   end
 
-  test "non existing is not rendered" do
-    code =
-      quote do
-        ~F"""
-        <#NonExisting />
-        """
-      end
+  if Version.match?(System.version(), ">= 1.15.0") do
+    test "non existing is not rendered" do
+      code =
+        quote do
+          ~F"""
+          <#NonExisting />
+          """
+        end
 
-    output =
-      capture_io(:standard_error, fn ->
-        assert_raise(CompileError, ~r/code:1: module NonExisting is not loaded and could not be found/, fn ->
-          compile_surface(code)
+      diagnostics =
+        Code.with_diagnostics(fn ->
+          try do
+            compile_surface(code)
+          rescue
+            e -> e
+          end
         end)
-      end)
 
-    assert output =~ ~r/cannot render <#NonExisting> \(module NonExisting could not be loaded\)/
+      assert {%CompileError{},
+              [
+                %{message: "cannot render <#NonExisting> (module NonExisting could not be loaded)" <> _},
+                %{message: "module NonExisting is not loaded and could not be found", severity: :error}
+              ]} = diagnostics
+    end
+  else
+    # Remove this test (and the `if`) whenever we drop support for Elixir < 1.15
+    test "non existing is not rendered" do
+      code =
+        quote do
+          ~F"""
+          <#NonExisting />
+          """
+        end
+
+      output =
+        capture_io(:standard_error, fn ->
+          assert_raise(CompileError, ~r/code:1: module NonExisting is not loaded and could not be found/, fn ->
+            compile_surface(code)
+          end)
+        end)
+
+      assert output =~ ~r/cannot render <#NonExisting> \(module NonExisting could not be loaded\)/
+    end
   end
 end